nbest_size can be negative

+++ include/onmt/SentencePiece.h
@@ -14,14 +14,14 @@ namespace onmt
   public:
     SentencePiece(const std::string& model_path);
 
-    void enable_regularization(size_t nbest_size, double alpha);
+    void enable_regularization(int nbest_size, double alpha);
 
     std::vector<std::string> encode(const std::string& str) const override;
     std::vector<AnnotatedToken> encode_and_annotate(const AnnotatedToken& token) const override;
 
   private:
     sentencepiece::SentencePieceProcessor _processor;
-    size_t _nbest_size;
+    int _nbest_size;
     double _alpha;
   };
